+++ src/main/java/com/felypeganzert/cacapalavras/repository/LetraRepository.java
@@ -4,6 +4,6 @@ import com.felypeganzert.cacapalavras.entidades.Letra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LetraRepository extends JpaRepository<Letra, Long>{
+public interface LetraRepository extends JpaRepository<Letra, Integer>{
     
 }
