+++ src/main/java/com/felypeganzert/cacapalavras/repository/LocalizacaoLetraNoTabuleiroRepository.java
@@ -4,6 +4,6 @@ import com.felypeganzert.cacapalavras.entidades.LocalizacaoLetraNoTabuleiro;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalizacaoLetraNoTabuleiroRepository extends JpaRepository<LocalizacaoLetraNoTabuleiro, Long>{
+public interface LocalizacaoLetraNoTabuleiroRepository extends JpaRepository<LocalizacaoLetraNoTabuleiro, Integer>{
     
 }
