+++ src/main/java/com/felypeganzert/cacapalavras/repository/LocalizacaoPalavraNoTabuleiroRepository.java
@@ -4,6 +4,6 @@ import com.felypeganzert.cacapalavras.entidades.LocalizacaoPalavraNoTabuleiro;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalizacaoPalavraNoTabuleiroRepository extends JpaRepository<LocalizacaoPalavraNoTabuleiro, Long>{
+public interface LocalizacaoPalavraNoTabuleiroRepository extends JpaRepository<LocalizacaoPalavraNoTabuleiro, Integer>{
     
 }
