+++ src/main/java/com/felypeganzert/cacapalavras/repository/TabuleiroRepository.java
@@ -4,6 +4,6 @@ import com.felypeganzert.cacapalavras.entidades.Tabuleiro;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TabuleiroRepository extends JpaRepository<Tabuleiro, Long>{
+public interface TabuleiroRepository extends JpaRepository<Tabuleiro, Integer>{
     
 }
